249 dm 완성하기 (#279)
* feat(backend): dm chat.controller

* feat(backend): get userInfo By nickname

* feat(backend): remove dm /request

* feat(backend): chat-create-dm event

* feat(backend): findDm

* feat(frontend): socket chat context

* feat(frontend): make socketChatCreateDmEvent

* feat: update channel list 로직, DM 만들고 작동 가능, leave 못하게 백엔드 로직 구현

* feat(chat): type 정리

* feat(frontend): socket context 파일 함수 wrapper하기

* feat(backend): app controller userInfo 가져오기

* feat(chat): block List 로직만 추가(테스트는 못해봄)

* feat(frontend): make Private Channel List component

* merge main

* wip: chat-join-dm

* fix: chatting event fix

이제 채팅 잘됨

* fix: dm 잘 보내짐

* fix: ui 개선

---------

+++ backend/src/chat/chat.gateway.ts
@@ -7,7 +7,7 @@ import { Gateway2faGuard } from 'src/guard/gateway2fa.guard';
 import { PnJwtPayload, PnPayloadDto } from 'src/dto/pnPayload.dto';
 import { JwtService } from '@nestjs/jwt';
 import { UserService } from 'src/user.service';
-import { IntraId } from 'src/type/userType';
+import { IntraId, Nickname } from 'src/type/userType';
 
 @WebSocketGateway({
   cors: { origin: '*' },
@@ -109,6 +109,7 @@ export class ChatGateway {
     const userInfo = this.userService.checkChatClient(client.id, pnPayload.intraId);
     if (!userInfo) return ;
     const channel = this.chatService.getChannel(channelId);
+    if (channel.channelType === 'private') return ;
     client.leave(channelId);
     this.chatService.leaveChannel(channelId, pnPayload.intraId);
     this.userService.deleteUserInfoChatRoomList(pnPayload.intraId, channelId);
@@ -317,6 +318,7 @@ export class ChatGateway {
         chatRoomList: [],
         gameRoom: '',
         online: true,
+        blockList: [],
       });
       return ;
     } else {
@@ -360,5 +362,42 @@ export class ChatGateway {
     this.server.to(channel.id).emit('chat-response-channel-info', { channel });
     return true;
   }
+
+  // 소매 넣기
+  @SubscribeMessage('chat-create-dm') // create = make + join
+  async handleCreateDm(@ConnectedSocket() client: Socket, @MessageBody() payloadEmit: { nickname: Nickname }, @PnJwtPayload() pnPayload: PnPayloadDto) {
+    const userInfo = this.userService.checkChatClient(client.id, pnPayload.intraId);
+    if (!userInfo) return ;
+
+    const title = pnPayload.nickname + ':' + payloadEmit.nickname;
+    const channel = this.chatService.findDm(pnPayload.nickname, payloadEmit.nickname);
+
+    /* 이미 dm이 존재하는 경우 */
+    console.log('chat-create-dm, channel', channel);
+    if (channel) {
+      this.server.to(userInfo.client.chat.id).emit('chat-join-dm', { channel });
+      return ;
+    }
+
+    /* dm이 존재하지 않는 경우 */
+    const channelId = this.chatService.addChannel(
+      { title, channelType: 'private', maxUsers: 2 },
+      client,
+      pnPayload.intraId,
+      pnPayload.nickname,
+    );
+    const newChannel = this.chatService.getChannel(channelId);
+    const opponentUser = await this.userService.getUserInfoByNickname(payloadEmit.nickname);
+    const opponentUserInfo = this.userService.getUserInfo(opponentUser.intraId);
+
+    client.join(newChannel.id);
+    opponentUserInfo.client.chat.join(newChannel.id);
+    this.chatService.joinChannel(channelId, pnPayload.intraId, pnPayload.nickname);
+    this.chatService.joinChannel(channelId, opponentUser.intraId, opponentUser.nickname);
+    this.syncAfterChannelChange(newChannel);
+    this.syncChannelList();
+    this.server.to(userInfo.client.chat.id).emit('chat-join-dm', { channel: newChannel });
+  }
+
 }
